NO-TICKET: Remove case sensitivity for Enum fromValue
There doesn't seem to be a good reason one would want fromValue to be case sensitive, but a case
for case insensitivty is because clients pass enums as url path params in a request, often with
no regard to case sensitivity.

In the strictest sense the client is in the wrong here as (urls ARE case sensitive. But in practice
most humans and clients don't care and probably shouldn't unless there is a good reason to do so.

Change-Id: I69ae5e3a4c1c380ab606f98ce09b40c2f8683b19